+++ CHANGELOG.md
@@ -18,6 +18,11 @@
   </summary>
 
 
+### :house: Internal
+- `scripts`
+  - [#2750](https://github.com/wix/ricos/pull/2750) updateBaselines - normalize keys in baselines 
+
+
 </details>
 <hr/>